Move deploy to admin
