+++ client/go/internal/admin/deploy/activate.go
