+++ client/go/internal/admin/deploy/curl.go
