+++ client/go/internal/admin/deploy/fetch.go
