+++ client/go/internal/admin/deploy/persist.go
