+++ client/go/internal/admin/deploy/prepare.go
