+++ client/go/internal/admin/deploy/results.go
