+++ client/go/internal/admin/deploy/upload.go
